# Fix two issues from #108