--- a/web/weaveHandler.go
+++ b/web/weaveHandler.go
@@ -23,13 +23,15 @@ const (
 	WEAVE_SIZE_LIMIT_EXCEEDED = "17" // Batch X-Weave-* headers too large
 )
 
-func WeaveInvalidWBOError(w http.ResponseWriter, r *http.Request) {
+func WeaveInvalidWBOError(w http.ResponseWriter, r *http.Request, reason error) {
+	logRequestProblem(r, http.StatusBadRequest, reason)
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusBadRequest)
 	w.Write([]byte(WEAVE_INVALID_WBO))
 }
 
-func WeaveSizeLimitExceeded(w http.ResponseWriter, r *http.Request) {
+func WeaveSizeLimitExceeded(w http.ResponseWriter, r *http.Request, reason error) {
+	logRequestProblem(r, http.StatusBadRequest, reason)
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusBadRequest)
 	w.Write([]byte(WEAVE_SIZE_LIMIT_EXCEEDED))