Found two large bugs in handling requests from iOS: 

1. `web/ReadNewlineJSON` has a bunch of bugs: 
  - it returned bad json raw data because of improper use of `scanner.Bytes()`. It didn't make a copy of the output. Since slices are references, the underlying data changed and inconsistent data came out
  - the scanner's default buffer, 64KB was too small. Increased to 257KB will fit all valid BSO payloads. The buffer being too small resulted in not actually finding the individual BSOs
2. Insufficient handling of values in `Content-Type` headers. We really only need the "mediatype" not all the params. Now it is extracted and checked correctly 

Notably, I took some time to optimize ReadNewlineJSON. By using a `sync.Pool` for its scanner's buffer it reduces memory allocation overhead and increases performance. The benchcmp output: 

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkReadNewlineJSON-8     112636        95069         -15.60%

benchmark                      old allocs     new allocs     delta
BenchmarkReadNewlineJSON-8     26             26             +0.00%

benchmark                      old bytes     new bytes     delta
BenchmarkReadNewlineJSON-8     603912        334257        -44.65%
```